tests/wpt/tests/navigation-api/navigate-event/navigate-form-with-target.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="iframe" name="i" src="/common/blank.html"></iframe>
<form id="form" method="post" action="foo.html" target="i"></form>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    navigation.onnavigate = t.unreached_func("onnavigate should not have fired in source window");

    iframe.contentWindow.navigation.onnavigate = t.step_func_done(e => {
      assert_equals(e.navigationType, "push");
      assert_true(e.cancelable);
      assert_true(e.canIntercept);
      assert_false(e.userInitiated);
      assert_false(e.hashChange);
      assert_equals(e.downloadRequest, null);
      assert_equals(new URL(e.destination.url).pathname,
                    "/navigation-api/navigate-event/foo.html");
      assert_false(e.destination.sameDocument);
      assert_equals(e.destination.key, "");
      assert_equals(e.destination.id, "");
      assert_equals(e.destination.index, -1);
      assert_not_equals(e.formData, null);
      assert_equals(e.sourceElement, null);
    });
    form.submit();
  });
}, "<form> submission with a target fires navigate event in target window but not source");
</script>
