third_party/blink/web_tests/external/wpt/navigation-api/navigate-event/navigate-form-with-target-expected.txt

This is a testharness.js-based test.
[FAIL] <form> submission with a target fires navigate event in target window but not source
  assert_equals: expected Element node <form id="form" method="post" action="foo.html" target="i... but got null
Harness: the test ran to completion.
